ScalarEnlistment: undo partial classes
The GVFS/GVFS.Common/GVFSEnlistment.cs file was split into two
and GVFS/GVFS.Common/GVFSEnlistment.Shared.cs was created in
commit ed0479021d547cb8d79905e5b510245879782f3c, in order to
share the IsUnattended() method with the GVFS.Hooks project.

Given that there is no longer a corresponding Scalar version of
the GVFS.Hooks project, there is no need to split out certain
(static) methods from the others in any classes.  (The GVFS.Hooks
project was removed in commit
f12effc371d7b1457ab58c6d44959632404e5998 in PR #23.)

Thus we merge the IsUnattended() method into ScalarEnlistment.cs
and remove the ScalarEnlistment.Shared.cs file.
