+++ Scalar.Common/ScalarEnlistment.cs
@@ -1,11 +1,13 @@
 using Scalar.Common.FileSystem;
 using Scalar.Common.Git;
+using Scalar.Common.Tracing;
 using System;
 using System.IO;
+using System.Security;
 
 namespace Scalar.Common
 {
-    public partial class ScalarEnlistment : Enlistment
+    public class ScalarEnlistment : Enlistment
     {
         public ScalarEnlistment(string enlistmentRoot, string workingDirectory, string repoUrl, string gitBinPath, GitAuthentication authentication)
             : base(
@@ -171,6 +173,26 @@ namespace Scalar.Common
             }
 
             return true;
+        }
+
+        public static bool IsUnattended(ITracer tracer)
+        {
+            try
+            {
+                return Environment.GetEnvironmentVariable(ScalarConstants.UnattendedEnvironmentVariable) == "1";
+            }
+            catch (SecurityException e)
+            {
+                if (tracer != null)
+                {
+                    EventMetadata metadata = new EventMetadata();
+                    metadata.Add("Area", nameof(ScalarEnlistment));
+                    metadata.Add("Exception", e.ToString());
+                    tracer.RelatedError(metadata, "Unable to read environment variable " + ScalarConstants.UnattendedEnvironmentVariable);
+                }
+
+                return false;
+            }
         }
     }
 }
